Bump v0.2.12; update Hytale API & region APIs
Release bump to 0.2.12 and update for Hytale Patch 5. Updates include:
- CHANGELOG and manifest version/serverVersion updated; build.gradle version bumped and Hytale server dependency changed to com.hypixel.hytale:Server:0.5.1.
- Deleted GEMINI.md and updated embedded AverageHytaleCore.jar.
- Migration of vector/rotation types to org.joml across models, services and tests (PlayerHome, PlayerRegionGroup, RegionZone, RegionService, RegionZoneTest).
- RegionService API changed (getRegionChunkAtPosition now uses Vector3i) and callers updated accordingly in ECS event handlers; event handlers now use playerRef for permission checks and messaging.
- PlayerJoinListener updated to obtain player from PlayerRef and use username for welcome messages; several unused imports removed.

These changes adapt the codebase to the newer Hytale API and adjust region/block interaction paths and player references accordingly.

+++ src/main/java/github/renderbr/hytale/db/models/regions/service/RegionService.java
@@ -1,6 +1,5 @@
 package github.renderbr.hytale.db.models.regions.service;
 
-import com.hypixel.hytale.math.vector.Vector3d;
 import com.hypixel.hytale.server.core.Message;
 import com.hypixel.hytale.server.core.entity.entities.Player;
 import com.hypixel.hytale.server.core.permissions.PermissionsModule;
@@ -13,6 +12,8 @@ import github.renderbr.hytale.db.models.regions.PlayerRegionGroup;
 import github.renderbr.hytale.db.models.regions.PlayerRegionGroupShare;
 import github.renderbr.hytale.db.models.regions.param.RegionZone;
 import github.renderbr.hytale.registries.ProviderRegistry;
+import org.joml.Vector3d;
+import org.joml.Vector3i;
 import util.ColorUtils;
 
 import javax.annotation.Nullable;
@@ -91,7 +92,7 @@ public class RegionService {
         return getIntersectingRegionsFromRect(RegionZone.getFromPosition(pos, 4), worldUuid);
     }
 
-    public PlayerRegionChunk getRegionChunkAtPosition(Vector3d pos, String worldUuid) throws SQLException {
+    public PlayerRegionChunk getRegionChunkAtPosition(Vector3i pos, String worldUuid) throws SQLException {
         return getRegionChunkTable().queryBuilder()
                 .where()
                 .le("firstCornerX", (int) pos.x).and().ge("secondCornerX", (int) pos.x).and()
